Extracts loading into a separate entity
