+++ testing/tests/DevExpress.ui.widgets.dataGrid/rowsView.tests.js
@@ -5552,7 +5552,7 @@ QUnit.module('Virtual scrolling', {
             rowsView._dataController._setViewportPositionCore = x._setViewportPositionCore;
             rowsView._dataController.option = rowsView.option.bind(rowsView);
             rowsView._dataController.positionChanged = $.Callbacks();
-            rowsView._dataController._dataSource = {
+            rowsView._dataController._dataOptions = {
                 changingDuration: function() { return 50; },
                 totalItemsCount: function() {
                     const virtualItemsCount = dataController.virtualItemsCount();
@@ -5581,6 +5581,7 @@ QUnit.module('Virtual scrolling', {
                 end: 7
             }
         };
+
         const dataController = new MockDataController(options);
         const rowsView = this.createRowsView(options.items, dataController);
         const testElement = $('#container');
